refactor: move agent_flow_engine.py out of experimental, rename to agentflow_engine.py
The unified engine has been the single execution surface for both eval
and training since #564. It's no longer experimental — promote it from
``rllm/experimental/engine/agent_flow_engine.py`` to
``rllm/engine/agentflow_engine.py`` (squashed module name to match the
``AgentFlowEngine`` / ``AgentFlow`` naming family already used elsewhere).

Test file renamed in lockstep
(``tests/engine/test_agent_flow_engine.py`` →
``tests/engine/test_agentflow_engine.py``).

All importers updated to ``rllm.engine.agentflow_engine``: ``rllm/eval/``
(_hooks, _resolution, materialize, runner), ``rllm/experimental/unified_trainer.py``,
``rllm/sandbox/sandboxed_flow.py``, ``rllm/tasks/{__init__,loader}.py``,
``rllm/types.py``, the engine test, the eval engine test, and the docs
page.

Drive-by fix: ``rllm/eval/runner.py``'s top-level imports of
``AgentFlowEngine`` + ``EvalGatewayManager`` triggered a circular import
when ``rllm.engine.agentflow_engine`` was imported cold (no other rllm
package loaded first): the engine pulls ``rllm.eval.types``, the eval
package init eagerly imports ``runner``, which imports back into the
half-loaded engine. The cycle was latent on main (pytest's collection
order avoided it) but a real ergonomic problem for cold imports. Fixed
by lazy-importing inside ``run_dataset``.

Verified: 423/423 unit tests pass; ``rllm eval math500 --agent
strands_math --max-examples 1`` returns 100%; cold ``from
rllm.engine.agentflow_engine import AgentFlowEngine`` works without a
prior import.
