+++ rllm/types.py
@@ -38,7 +38,7 @@ class Task:
 
     Pure data — describes itself (instruction, metadata) and points at the
     directory where its verifier lives.
-    :class:`rllm.experimental.engine.agent_flow_engine.AgentFlowEngine`
+    :class:`rllm.engine.agentflow_engine.AgentFlowEngine`
     (driven by :class:`rllm.eval._hooks.EvalHooks` at eval time) reads
     the task config and resolves the appropriate :class:`Evaluator` at
     run time.
