+++ tests/engine/test_agentflow_engine.py
@@ -1,8 +1,8 @@
 import asyncio
 
 from rllm.agents.agent import Episode, Trajectory
+from rllm.engine.agentflow_engine import AgentFlowEngine
 from rllm.eval.types import EvalOutput
-from rllm.experimental.engine.agent_flow_engine import AgentFlowEngine
 from rllm.workflows.workflow import TerminationReason
